build: add nix flake
Define the project build logic comprehensively, both as a convenience to
developers when getting started, and to serve as a reference for
operators who wish to build the source code their own way.

Theoretically the devshell should be cross-platform, but I've only
tested it on Linux hosts, not on macOS.

I'll continue to update the tooling refs to keep pace with the repo,
simply to keep my own development environment sane.